refactor: move and rename hooks adn helpers folders from DownloadAnShareComponents
