+++ src/components/TeacherComponents/LessonItemContainer/LessonItemContainer.tsx
@@ -1,7 +1,7 @@
 import { forwardRef } from "react";
 
 import { LessonPageLinkAnchorId } from "@/components/TeacherComponents/lesson.helpers";
-import { containerTitleToPreselectMap } from "@/components/DownloadAndShareComponents/helpers/containerTitleToPreselectMap";
+import { containerTitleToPreselectMap } from "@/components/TeacherComponents/downloadAndShareHelpers/containerTitleToPreselectMap";
 import { LessonItemContainerLink } from "@/components/TeacherComponents/LessonItemContainerLink";
 import { Heading, Hr } from "@/components/SharedComponents/Typography";
 import AnchorTarget from "@/components/SharedComponents/AnchorTarget";
