+++ src/components/TeacherComponents/downloadAndShareHelpers/createAndClickHiddenDownloadLink.tsx
