+++ src/components/TeacherComponents/downloadAndShareHelpers/createDownloadResourcesLink.tsx
