+++ src/components/TeacherComponents/downloadAndShareHelpers/getDownloadResourceType.test.tsx
