+++ src/components/TeacherComponents/downloadAndShareHelpers/getDownloadResourcesExistence.test.tsx
