+++ src/components/TeacherComponents/downloadAndShareHelpers/getDownloadResourcesExistence.tsx
@@ -1,7 +1,6 @@
 import { z } from "zod";
 
-import { DownloadResourceType } from "../downloadAndShare.types";
-
+import { DownloadResourceType } from "@/components/DownloadAndShareComponents/downloadAndShare.types";
 import OakError from "@/errors/OakError";
 import getBrowserConfig from "@/browser-lib/getBrowserConfig";
