+++ src/components/TeacherComponents/downloadAndShareHelpers/getFormattedDetailsForTracking.test.tsx
